dll to static lib and clean-up

+++ NeuralEngine/src/NeuralEngine/Platform/Windows/WinWindow.cpp
@@ -129,7 +129,7 @@ namespace NeuralEngine
 			{
 				// Retrieve window data and trigger a mouse scroll event
 				WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-				MouseScrolledEvent event(xoffset, yoffset);
+				MouseScrolledEvent event((float)xoffset, (float)yoffset);
 				data.callback(event);
 			});
